refactor: moving the attendance form to the cog
