+++ src/data.py
@@ -15,21 +15,6 @@ Functions:
 
 from datetime import datetime
 
-MONTHS = [
-    "janeiro",
-    "fevereiro",
-    "marco",
-    "abril",
-    "maio",
-    "junho",
-    "julho",
-    "agosto",
-    "setembro",
-    "outubro",
-    "novembro",
-    "dezembro",
-]
-
 
 def _row_to_project(row: str) -> dict:
     """
