+++ src/reports/attendance_sheet.py
@@ -22,8 +22,7 @@ from reportlab.lib.pagesizes import A4
 from reportlab.lib.units import cm
 from reportlab.platypus import Paragraph, SimpleDocTemplate, Spacer, Table, TableStyle
 
-from attendances_data import Attendance
-from data import MONTHS
+from attendances_data import MONTHS, Attendance
 
 from . import styles
 from .commons import setup_header
